Atualização (pré-banca)

+++ app/view/produto/listProd.php
@@ -9,7 +9,7 @@ require_once(__DIR__ . "/../../controller/produtoController.php");
 ?>
 
 <div class="container">
-    
+
     <div class="row">
         <div class="col-6">
             <?php require_once(__DIR__ . "/../include/msg.php"); ?>
@@ -23,6 +23,7 @@ require_once(__DIR__ . "/../../controller/produtoController.php");
                     <tr>
                         <th>ID</th>
                         <th>Nome</th>
+                        <th>Preço</th>
                         <th>Categorias</th>
                         <th>Detalhes</th>
                         <th>Alterar</th>
@@ -30,31 +31,31 @@ require_once(__DIR__ . "/../../controller/produtoController.php");
                     </tr>
                 </thead>
                 <tbody>
-                    <?php 
-                        foreach($dados['listaProd'] as $prod): 
+                    <?php
+                    foreach ($dados['listaProd'] as $prod) :
                     ?>
                         <tr>
                             <td><?php echo $prod->getIdProduto(); ?></td>
                             <td><?= $prod->getNomeProduto(); ?></td>
-                            <td><?= $prod->getCategoriaProduto(); ?></td>
+                            <td><?= $prod->getPrecoProduto(); ?></td>
+                            <td><?= $prod->getCategoriaDesc(); ?></td>
                             <td><?= $prod->getDetalhes(); ?></td>
                             <td><a class="btn btn-primary" href="../controller/produtoController.php?action=editProd&id=<?= $prod->getIdProduto() ?>">
-                                Alterar</a> 
+                                    Alterar</a>
                             </td>
-                            <td><a class="btn btn-danger" 
-                                onclick="return confirm('Confirma a exclusão do produto?');"
-                                href="<?= BASEURL ?>/controller/produtoController.php?action=deleteProd&id=<?= $prod->getIdProduto() ?>">
-                                Excluir</a> 
+                            <td><a class="btn btn-danger" onclick="return confirm('Confirma a exclusão do produto?');" href="<?= BASEURL ?>/controller/produtoController.php?action=deleteProd&id=<?= $prod->getIdProduto() ?>">
+                                    Excluir</a>
                             </td>
                         </tr>
                     <?php endforeach; ?>
                 </tbody>
             </table>
+            <a href="../controller/homeController.php?action=homeVendedor" class="btn btn-success ">Voltar</a>
         </div>
     </div>
 
 </div>
 
-<?php  
+<?php
 require_once(__DIR__ . "/../include/footer.php");
 ?>
